Added the code to support parsing of other keys from feeds
I have added a NSArray that is used to pass in a list of NSStrings that
is to contain the xml element names of the custom elements that you
would like MWFeedParser to look for and parse if it is found.  If the
elements are found they will be added to the customProperities
NSDictionary that I have added to the MWFeedItem class.