Add feedURL property to MWFeedInfo class
Generally speaking, it seems wise that since you can get a `MWFeedInfo` object from a
`MWFeedParser`, you should be able to reconstruct an `MWFeedParser` from an `MWFeedInfo`.
This pull request adds this functionality. Below is a common use case that demonstrates
this.

In a situation slightly more advanced than the provided demo, but a common scenario worth
accommodating in the library, you might have a list of several feeds, all of which
require their own `MWFeedParser`. If the list of `MWFeedParser` objects are in a table
view, and a particular feed is selected, you would likely pass the `MWFeedInfo` object
into the next view controller to display the feed items. However, `MWFeedInfo` does not
currently keep track of the URL it was retrieved from, which requires keeping track of it
separately or passing in the parser, neither of which are ideal. This pull request adds
a `feedURL` property to the `MWFeedInfo` object, allowing a view controller to construct
its own `MWFeedParser` with its own settings to retrieve the items.