separate categories for NSDate/NSDateFormatter/NSString+InternetDateTime
there are a lot of times i have a string that is in internet date-time
format that i simply want a readable string; creating a category off of
NSString seems like the way to go for this, and since that code is now
string based, and to avoid duplication, just re-factored the code out of
NSDate into NSString where appropriate.

once that was done, NSDateFormatter was needed in both, so for the two
singletons it contained, just create a separate category to access them.